+++ tests/OptionOverriding/Input/test.xpybuild.py
@@ -1,4 +1,4 @@
-import os
+import os, logging
 from propertysupport import *
 from buildcommon import *
 from pathsets import *
@@ -6,14 +6,10 @@ from pathsets import *
 defineOutputDirProperty('OUTPUT_DIR', None)
 
 import basetarget 
-defineOption('mydefopt', 'myval')
-defineOption('myopt', 'orig-shouldnotsee')
-defineOption('staticopt', 'orig')
-defineOption('staticopt2', 'orig')
-defineOption('dynopt', 'orig')
-setGlobalOption('myopt', 'myval')
 
-defineOption('testoption.default', 'expectedval')
+defineStringProperty('MY_PROP', 'expectedval')
+
+defineOption('testoption.default', '${MY_PROP}')
 defineOption('testoption.globalOverride', 'unexpectedval')
 setGlobalOption('testoption.globalOverride', 'expectedval')
 
@@ -21,13 +17,33 @@ defineOption('testoption.targetOverride', 'defaultval')
 
 defineOption('testoption.legacyTargetOverride', 'defaultval')
 
+defineOption('testoption2.empty', '')
+
+class CustomPathSet(PathSet):
+	def __init__(self, *args, **kwargs):
+		PathSet.__init__(self, *args, **kwargs)
+	def resolveWithDestinations(self, context):
+		assert self.target.options
+		return PathSet.resolveWithDestinations(self, context)
+	def _resolveUnderlyingDependencies(self, context):
+		# this might be needed in some targets, so check it works
+		assert self.target.options
+		logging.getLogger('custompathset').critical('PathSet._resolveUnderlyingDependencies got options: %s', self.target.options)
+		return PathSet._resolveUnderlyingDependencies(self, context)
 
 class MyTarget(basetarget.BaseTarget):
-	def __init__(self, name, options=None):
-		super(MyTarget, self).__init__(name, [])
+	def __init__(self, name, options=None, deps=[]):
+		super(MyTarget, self).__init__(name, dependencies=deps)
+		
 		if options:
 			# this is no longer recommended practice, but need to maintain it for backwards compat
 			self.options = options
+
+		try:
+			x = self.options
+			self.log.error('ERROR - should be impossible to read self.options in the constructor unexpectedly: %s', x)
+		except Exception, e:
+			self.log.critical('Exception from trying to read self.options: %s'%e)
 		
 	def run(self, context):
 		name = os.path.basename(self.path)
@@ -39,11 +55,29 @@ class MyTarget(basetarget.BaseTarget):
 						self.log.critical('-- %s %s=%s', display, k, o[k])
 						print >>f, '%s %s=%s'%(display, k, o[k])
 		
-			printit(self.options, name+' self.options')
 			printit(context.mergeOptions(self), name+' mergeOptions')
+			assert context.mergeOptions(self) == self.options, '%s != %s'%(context.mergeOptions(self), self.options)
+			self.log.critical('-- %s getOption %s=%s', name, 'testoption.default', self.getOption('testoption.default'))
+			
+			try:
+				self.getOption('testoption2.empty')
+				self.log.error('ERROR - getOption should throw if empty')
+			except Exception, e:
+				self.log.critical('Got expected exception: %s', e)
+			
+			
+	def getHashableImplicitInputs(self, context):
+		assert self.options
+		return [str(self.options)]
 		
-	def clean(self, context): pass
+	def clean(self, context): 
+		pass
 
-MyTarget('${OUTPUT_DIR}/defaults.txt')
-MyTarget('${OUTPUT_DIR}/legacyTargetOverride.txt', options={'testoption.legacyTargetOverride':'expectedval'})
+pathset = CustomPathSet('test.xpybuild.py')
+t = MyTarget('${OUTPUT_DIR}/defaults.txt', deps=pathset)
+pathset.target = t
+
+MyTarget('${OUTPUT_DIR}/legacyTargetOverride.txt', options={'testoption.legacyTargetOverride':'${MY_PROP}'})
 MyTarget('${OUTPUT_DIR}/targetOverride.txt').option('testoption.targetOverride', 'expectedval')
+# do both and check .option(...) takes precedence
+MyTarget('${OUTPUT_DIR}/targetOverrideBoth.txt', options={'testoption.targetOverride':'unexpectedval'}).option('testoption.targetOverride', 'expectedval')